requirements.txt
openai_whisper==20230314
pandas==1.5.3
streamlit==1.24.1
streamlit_scrollable_textbox==0.0.3
